adding Spring Security

+++ secondBookStore/src/main/java/fi/haagahelia/secondBookStore/domain/Book.java
@@ -81,7 +81,7 @@ public class Book {
 		return price;
 	}
 
-	public void setPrice(int price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
 	public Category getCategory() {
